增加更新集群信息接口
