+++ src/kube/types/types.go
@@ -314,8 +314,8 @@ const (
 )
 
 const (
-	// FolderID 每个cluster只有唯一的 folder，此节点没有表，统一用 999 表示 folder的ID，如果需要确认具体的 folder 需要与clusterID
-	// 结合起来使用
+	// KubeFolderID 每个cluster只有唯一的 folder，此节点没有表，统一用 999 表示 folder的ID，如果需要确认具体的 folder
+	// 需要与clusterID结合起来使用
 	KubeFolderID   = 999
 	KubeFolderName = "空Pod节点"
 )
@@ -346,7 +346,7 @@ const (
 	// LimitsField container limits field
 	LimitsField = "limits"
 
-	// container requests field
+	// RequestsField requests field
 	RequestsField = "requests"
 
 	// LivenessField container liveness field
