+++ src/source_controller/coreservice/core/core.go
@@ -124,7 +124,7 @@ type InstanceOperation interface {
 type ContainerOperation interface {
 	CreateCluster(kit *rest.Kit, bizID int64, option *types.ClusterBaseFields) (int64, errors.CCErrorCoder)
 	UpdateClusterFields(kit *rest.Kit, bizID int64, supplierAccount string,
-		data *types.UpdateClusterOption) errors.CCErrorCoder
+		data *types.UpdateClusterOption) (*metadata.UpdatedCount, errors.CCErrorCoder)
 	SearchCluster(kit *rest.Kit, input *metadata.QueryCondition) (*types.ResponseCluster, error)
 	CreatePod(kit *rest.Kit, bizID int64, data *types.PodCoreInfo) (int64, errors.CCErrorCoder)
 	CreateContainer(kit *rest.Kit, bizID int64, data *types.ContainerCoreInfo) (int64,
